Add files via upload
I made a project based on rock-paper-scissors game in Java. It can be played in two ways as User - Computer or Computer-Computer. The user determines his choice in each round, and the computer chooses randomly. When the end conditions of the game are met, scores are calculated according to certain rules and the winner is printed. All these processes designed in a visual interface.